requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
python-ldap==3.2.0
